Implement MergeSort Threading Proof of Concept
Initially, the threaded merge sort tried to directly sort the
comparables array. This could be causing cache missing. Two new arrays
are allocated and given one half of comparables each. Cache missing is
solved. Speed is halved from the un-threaded implementation. This is a
proof of concept, for it only uses two threads.