Implement Intelligent MergeSort Multi-Threading
Building on the prototype, I knew the merge-sort could benefit from
utilizing an increased number of threads. To make the program scale with
any number of threads/cores, it had to create threads with respect to
the number of threads available on the systems processor. I believe this
program will scale well on any processor. From a Dual-Core to a
ThreadRipper.